Patchy: write PID in config file, not in build dir
test-master-lock is specific to the main Git repository used, whereas
PID file was specific to the build directory associated to the branch
being built, so locking handling was half-broken.  As the Git
repository is defined in the config file, it is sensible to write the
PID in this same place.