Patchy: support running build commands with another user
For this to work non-interactively, sudo should be installed, and a
line like

gituser  ALL = (builduser) NOPASSWD: ALL

should be added to /etc/sudoers, where gituser is the user that runs
Patchy and builduser is the user that runs build commands (autogen.sh,
configure, make...). This avoids exposing Git push access and other
files and services to the build process.

Some additional features, like the installation of documentation,
build logs and regression tests comparison for serving them on the
WWW, may require in addition that gituser be a member of
lilybuild.