Patchy: install documentation if revision already pushed
Rename WastedBuildException to DuplicateBuildException, which is
actually raised when the revision that has just been built is exactly
the one of remote master; in this case, try to install documentation,
because the Patchy that has merged master may not have uploaded
documentation to some WWW location already.

Also add an option for advertising documentation URL in log.