Add script to generate lilypond-flow images
bash script that runs pdflatex on the source and then
converts the pdf to tow png files.

a translation suffix (e.g. "-de") can be passed as parameter.
Requires pdflatex (with used packages) and imagemagick
to be present, but I think this can be assumed anyway
as requirements for "build-dep lilypond"